dash-tui: refactor our project code
